Moved utils to utils folder
