+++ src/js/shot-generator-new/components/InspectedElement/HandInspector/HandPresetsEditor/HandPresetsEditorItem.js
@@ -2,7 +2,7 @@ import fs from 'fs-extra'
 import path from 'path'
 import { remote } from 'electron'
 import classNames from 'classnames'
-import ThumbnailRenderer from '../../../../ThumbnailRenderer'
+import ThumbnailRenderer from '../../../../utils/ThumbnailRenderer'
 import React, { useMemo } from 'react'
 import { createdMirroredHand, applyChangesToSkeleton, getOppositeHandName } from '../../../../../utils/handSkeletonUtils'
 import { GUTTER_SIZE, ITEM_WIDTH, ITEM_HEIGHT, IMAGE_HEIGHT, IMAGE_WIDTH } from '../../../../utils/InspectorElementsSettings'
