+++ src/js/shot-generator-new/components/InspectedElement/PosePresetsInspector/PosePresetInspectorItem.js
@@ -3,7 +3,7 @@ import React from 'react'
 import fs from 'fs-extra'
 import classNames from 'classnames'
 import '../../../../vendor/three/examples/js/utils/SkeletonUtils'
-import ThumbnailRenderer from '../../../ThumbnailRenderer'
+import ThumbnailRenderer from '../../../utils/ThumbnailRenderer'
 import { GUTTER_SIZE, ITEM_WIDTH, ITEM_HEIGHT, IMAGE_HEIGHT, IMAGE_WIDTH } from '../../../utils/InspectorElementsSettings'
 import { remote } from 'electron'
 import { useMemo } from 'react'
