+++ src/js/shot-generator-new/utils/apply-device-quaternion.js
@@ -39,4 +39,4 @@ const applyDeviceQuaternion = ({
   return objectQuaternion
 }
 
-module.exports = applyDeviceQuaternion
+export default applyDeviceQuaternion
